feat: dag robbery added
