+++ airflow/dags/dag_car_theft.py
@@ -1,0 +1,43 @@
+from datetime import datetime
+from airflow import DAG
+from airflow.operators.python import PythonOperator
+from airflow.providers.amazon.aws.transfers.local_to_s3 import LocalFilesystemToS3Operator
+from datetime import datetime
+from airflow.models import Variable
+import os
+from models.setup_car_threft import start_setup_car_threft
+
+default_args = {
+    "start_date": datetime(2003, 1, 1), 
+}
+
+def delete_file(name):
+    os.remove(name)
+
+with DAG(
+    "ssp_extraction_to_s3_car_theft", schedule="@monthly", catchup=True, max_active_runs=2, default_args=default_args
+) as dag:
+    filename = "{{ ds }}.xlsx"
+    execution_date = "{{ ds }}"
+    partition = "{{ macros.ds_format(ds, '%Y-%m-%d', 'year=%Y/month=%m') }}"
+    bucket = Variable.get("bucket_raw_default", "tcc_raw")
+
+    extract_task = PythonOperator(
+        task_id="extract", python_callable=start_setup_car_threft, 
+        op_kwargs={"reference_date": execution_date, "default_filename": filename}
+    )
+
+    load_task = LocalFilesystemToS3Operator(
+        task_id="load", 
+        filename=filename, 
+        dest_key=f"s3://{bucket}/car_theft/{partition}/{filename}",
+        replace=True
+    )
+
+    cleanup = PythonOperator(
+        task_id="cleanup",
+        python_callable=delete_file,
+        op_kwargs={"name": filename}
+    )
+
+    extract_task >> load_task >> cleanup
