+++ airflow/dags/models/config_ssp.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+BUTTON_PREFIX = "ctl00$cphBody${}"
+
+def button(id):
+    return BUTTON_PREFIX.format(id)
+
+class CrimeType(Enum):
+    CAR_THEFT=button("btnFurtoVeiculo")
+    CAR_ROBBERY=button("btnRouboVeiculo")
+
